data-pipeline/requirements.txt
# Data Pipeline Requirements
# For RAG pipeline: crawler -> chunker -> embedder -> qdrant

# Web scraping and content extraction
requests>=2.32.0
beautifulsoup4>=4.12.3
newspaper3k>=0.2.8
lxml>=5.3.0
certifi>=2024.7.4
lxml-html-clean>=0.2.0

# Text processing and embeddings
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

# Vector database
qdrant-client>=1.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# LLMs / LangChain
langchain>=0.2.3
langchain-core>=0.2.4
langchain-openai>=0.1.7
langchain-anthropic>=0.1.14
langchain-google-genai>=2.0.7
openai>=1.35.0
anthropic>=0.34.0
google-generativeai>=0.8.3
langchain-google-genai
